# Various performance optimizations and bug fixes