--- a/aggregation/timer.go
+++ b/aggregation/timer.go
@@ -36,16 +36,16 @@ type Timer struct {
 
 // NewTimer creates a new timer
 func NewTimer(opts cm.Options) Timer {
-	return Timer{
-		stream: opts.StreamPool().Get(),
-	}
+	stream := opts.StreamPool().Get()
+	stream.Reset()
+	return Timer{stream: stream}
 }
 
 // Add adds a timer value
 func (t *Timer) Add(value float64) {
 	t.count++
 	t.sum += value
-	t.sumSq += math.Pow(value, 2)
+	t.sumSq += value * value
 	t.stream.Add(value)
 }
 
@@ -93,7 +93,7 @@ func (t *Timer) Mean() float64 {
 
 // Stdev returns the standard deviation timer value
 func (t *Timer) Stdev() float64 {
-	num := float64(t.count)*t.sumSq - math.Pow(t.sum, 2)
+	num := float64(t.count)*t.sumSq - t.sum*t.sum
 	div := t.count * (t.count - 1)
 	if div == 0 {
 		return 0.0