--- a/server/msgpack/server.go
+++ b/server/msgpack/server.go
@@ -36,6 +36,10 @@ import (
 	"github.com/uber-go/tally"
 )
 
+const (
+	unknownRemoteHostAddress = "<unknown>"
+)
+
 type serverMetrics struct {
 	openConnections tally.Gauge
 	addMetricErrors tally.Counter
@@ -205,7 +209,7 @@ func (s *server) handleConnection(conn net.Conn) {
 	// If there is an error during decoding, it's likely due to a broken connection
 	// and therefore we ignore the EOF error
 	if err := it.Err(); err != nil && err != io.EOF {
-		remoteAddress := "<unknown>"
+		remoteAddress := unknownRemoteHostAddress
 		if remoteAddr := conn.RemoteAddr(); remoteAddr != nil {
 			remoteAddress = remoteAddr.String()
 		}