cc @robskillington @cw9 @prateek @jeromefroe 

This PR contains various performance optimizations for the aggregation tier. Specifically, it includes
* Adding an option to provide a nil sample pool, which effectively disables sample pooling. The sample pool is used to provide sample objects for every timer value, which is on the critical path and is heavily contended. Profile shows getting values from the pool and putting values into the pool accounts for >30% of the CPU due to heavy contention. Disabling sample pooling significantly reduced write p99 (4ms -> 1ms) and flush p99 (1.5s -> 200ms).
* Adding an option to periodically flush streams at write time, therefore reducing processing load during global flushing. This has significantly improved flush p99 for aggregation servers that have large batches of timer values. In one case, we saw 5s -> 150ms reduction w.r.t. p99 flush latency.
* Replace math.Pow(x, 2) with x * x. This used to account for 4% of the CPU due to math.Pow(x, 2) being 100x more expensive than the plain multiplication. 
* Preallocate enough buffer for the buffered encoder. The buffered encode is used to hold the aggregated metrics. Previous it starts with an empty buffer and gradually grows to the maximum flush size, which requires unnecessary copy and reallocations.

It also includes more instrumentation and bug fixes. Specifically, it includes:
* Fixing a memory leak in timer creation where the underlying stream is not reset, causing the stream (and the float slices and the samples) to not be properly returned to pool.
* Adding logging to record the remote address when a decode error happens. This has allowed us to determine the primary cause of the decode errors is due to misconfigured hosts in the wrong datacenter.